pyproject.toml
[project]
name = "mcp-py-venus-step"
version = "0.1.0"
authors = [{ name = "Bradley Bare", email = "bradleybare@gmail.com" }]
requires-python = "==3.13.*"
dependencies = [
    "peewee>=4.0.4,<5",
    "fastmcp>=3.2.3,<4",
    "pydantic>=2.12.5,<3", 
    "python-dotenv>=1.2.2,<2", 
    "py-venus-step @ git+https://github.com/BirdBare/py-venus-step",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_py_venus_step"]

[tool.uv]
package = true

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.dependencies]
uv = ">=0.11.6,<0.12"

[tool.pixi.tasks]
executor_server = {cmd="uv run executor.py", cwd="src/mcp_py_venus_step/shared", env={VENUS_CONNECTION_MODE="SOCKET", VENUS_CONNECTION_PATH="/Users/birdbare/Parallels/venus_star.sock", VENUS_EXECUTOR_PORT="57000"}}
